requirements.txt
Django
django-webpack-loader==0.7.0
xlrd3
django-formtools==2.3
django-simple-captcha
djangorestframework
gurobipy
pandas
